Moved vue translations to shared util
