+++ contentcuration/contentcuration/static/js/edit_channel/import/views.js
@@ -6,46 +6,9 @@ var BaseViews = require("../views");
 require("import.less"); // eslint-disable-line
 var dialog = require("../utils/dialog");
 var store = require('./vuex/store');
-var vueIntl = require("vue-intl");
-var translations = require("utils/translations");
 var { PageTypes } = require('./constants');
 
 var ImportModal = Vue.extend(ImportModalComponent);
-
-// Flatten translation dictionary
-var unnested_translations = {};
-Object.keys(translations).forEach(function (key) {
-    Object.keys(translations[key]).forEach(function(nestedKey) {
-        unnested_translations[key + "." + nestedKey] = translations[key][nestedKey];
-    });
-});
-
-Vue.use(vueIntl, {"defaultLocale": "en"});
-
-var currentLanguage = "en";
-if (global.languageCode) {
-    currentLanguage = global.languageCode;
-    Vue.setLocale(currentLanguage);
-}
-
-Vue.registerMessages(currentLanguage, unnested_translations);
-Vue.prototype.$tr = function $tr(messageId, args) {
-    const nameSpace = this.$options.name;
-    if (args) {
-        if (!Array.isArray(args) && typeof args !== 'object') {
-            logging.error(`The $tr functions take either an array of positional
-                            arguments or an object of named options.`);
-        }
-    }
-    const defaultMessageText = this.$options.$trs[messageId];
-    const message = {
-        id: `${nameSpace}.${messageId}`,
-        defaultMessage: defaultMessageText,
-    };
-
-    return this.$formatMessage(message, args);
-};
-
 
 function getImportStatus(state) {
   return state.import.importStatus;
